pyproject.toml
[build-system]
requires = ['setuptools>=61.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'mspec'
version = '0.0.1'
authors = [
  { name='Brad Corlett', email='brad.corlett@gmail.com' },
]
description = 'the medium protocol and stuff'
readme = 'README.md'
requires-python = '>=3.7'
dependencies = [
    'pyyaml == 6.0.2',
    'Jinja2 == 3.1.4'
]
